+++ src/fixed/module.rs
@@ -17,9 +17,9 @@ pub type UntargetedIdentifier = String;
 /// Effects in the order they get applied
 pub enum Effect {
     Capacitor(i16),
-    ArmorRepair(u32),
-    Damage(u32),
-    Mine(u32),
+    ArmorRepair(u16),
+    Damage(u16),
+    Mine(u16),
     WarpDisruption,
 }
 
@@ -27,21 +27,21 @@ pub enum Effect {
 #[cfg_attr(test, derive(ts_rs::TS))]
 #[serde(rename_all = "camelCase", rename = "ModulePassive")]
 pub struct Passive {
-    pub required_cpu: u32,
-    pub required_powergrid: u32,
+    pub required_cpu: u16,
+    pub required_powergrid: u16,
 
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub capacitor: Option<u32>,
+    pub capacitor: Option<u16>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub hitpoints_armor: Option<u32>,
+    pub hitpoints_armor: Option<u16>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 #[cfg_attr(test, derive(ts_rs::TS))]
 #[serde(rename_all = "camelCase", rename = "ModuleUntargeted")]
 pub struct Untargeted {
-    pub required_cpu: u32,
-    pub required_powergrid: u32,
+    pub required_cpu: u16,
+    pub required_powergrid: u16,
 
     #[serde(serialize_with = "ordered_vec")]
     pub effects: Vec<Effect>,
@@ -51,8 +51,8 @@ pub struct Untargeted {
 #[cfg_attr(test, derive(ts_rs::TS))]
 #[serde(rename_all = "camelCase", rename = "ModuleTargeted")]
 pub struct Targeted {
-    pub required_cpu: u32,
-    pub required_powergrid: u32,
+    pub required_cpu: u16,
+    pub required_powergrid: u16,
 
     #[serde(serialize_with = "ordered_vec")]
     pub effects_origin: Vec<Effect>,
